backend_python/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Financial Data Fetchers
akshare==1.13.3
yfinance==0.2.35

# Metrics Calculation
scipy==1.11.4
statsmodels==0.14.1

# Visualization
matplotlib==3.8.2
plotly==5.18.0
kaleido==0.2.1

# AI/LLM
google-generativeai==0.3.2
openai==1.10.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1

# Date & Time
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0
